requirements.txt
streamlit>=1.28.0
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
ffmpeg-python>=0.2.0
gspread>=5.0.0
pandas>=1.5.0
google-auth>=2.0.0 